+++ datadriven/src/sgpp/datadriven/operation/hash/OperationMPI/OperationDensityMultMPI.hpp
@@ -18,13 +18,13 @@ class OperationDensityMultMPI : public base::OperationMatrix, public DensityWork
     double *alpha_ptr = alphavector.getPointer();
     this->alpha = std::vector<double>(alpha_ptr, alpha_ptr+alphavector.size());
     send_alpha();
-    int datainfo[2];
+    long datainfo[2];
     datainfo[0] = 0;
     datainfo[1] = static_cast<int>(result.getSize());
     divide_workpackages(datainfo, result.getPointer());
-    int exitmessage[2] = {-2, -2};
+    long exitmessage[2] = {-2, -2};
     for (int dest = 1; dest < MPIEnviroment::get_sub_worker_count() + 1; dest++)
-      MPI_Send(exitmessage, 2, MPI_INT, dest, 1, sub_worker_comm);
+      MPI_Send(exitmessage, 2, MPI_LONG, dest, 1, sub_worker_comm);
   }
 };
