+++ datadriven/src/sgpp/datadriven/operation/hash/OperationMPI/OperationDensityRhsMPI.hpp
@@ -15,13 +15,13 @@ class OperationDensityRhsMPI : public DensityRhsWorker {
   }
   virtual void generate_b(base::DataVector &b) {
     DensityRhsWorker::MPIWorkerGridBase::start_sub_workers();
-    int datainfo[2];
+    long datainfo[2];
     datainfo[0] = 0;
     datainfo[1] = static_cast<int>(b.getSize());
     divide_workpackages(datainfo, b.getPointer());
-    int exitmessage[2] = {-2, -2};
+    long exitmessage[2] = {-2, -2};
     for (int dest = 1; dest < MPIEnviroment::get_sub_worker_count() + 1; dest++)
-      MPI_Send(exitmessage, 2, MPI_INT, dest, 1, sub_worker_comm);
+      MPI_Send(exitmessage, 2, MPI_LONG, dest, 1, sub_worker_comm);
   }
 };
